+++ frontend/src/app/features/work-packages/components/back-routing/back-routing.service.ts
@@ -49,7 +49,7 @@ export class BackRoutingService {
   constructor(readonly injector:Injector) {
   }
 
-  public goToOtherState(route:string, params:{}):Promise<any> {
+  public goToOtherState(route:string, params:Record<string, unknown>):Promise<unknown> {
     return this.$state.go(route, params);
   }
