+++ frontend/src/app/features/work-packages/routing/partitioned-query-space-page/partitioned-query-space-page.component.ts
@@ -29,7 +29,6 @@
 import {
   ChangeDetectionStrategy,
   Component,
-  ComponentRef,
   OnDestroy,
   OnInit,
 } from '@angular/core';
