+++ frontend/src/app/shared/components/principal/principal-list.component.ts
@@ -2,6 +2,7 @@ import {
   ChangeDetectionStrategy,
   Component,
   Input,
+  HostBinding,
 } from '@angular/core';
 import { PrincipalLike } from 'core-app/shared/components/principal/principal-types';
 
@@ -10,12 +11,9 @@ import { PrincipalLike } from 'core-app/shared/components/principal/principal-ty
   templateUrl: './principal-list.component.html',
   styleUrls: [],
   changeDetection: ChangeDetectionStrategy.OnPush,
-  host: {
-    class: 'op-principal-list',
-  },
 })
 export class OpPrincipalListComponent {
-  @Input() principals:PrincipalLike[];
+  @HostBinding('class.op-principal-list') className = true;
 
-  constructor() {}
+  @Input() principals:PrincipalLike[];
 }
